pyproject.toml
[project]
name = "claude-harness"
version = "0.1.0"
description = "Lightweight fleet harness for Claude Code — memory, messaging, self-evolution across machines."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "watchfiles>=0.21",
]

[project.scripts]
harness = "harness.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["harness*"]
exclude = ["repos*", "knowledge*", "docs*", "principles*", "dashboard*", "skill*", "hooks*", "roles*"]
